sam-todaylaw/TL_detail/requirements.txt
requests
boto3
pymongo
PyJWT
bs4